requirements.txt
# Image Processing
numpy>=1.26.0
opencv-python>=4.8.0.76
Pillow>=10.0.0
rawpy>=0.18.1
tifffile>=2023.7.10
piexif>=1.1.3

# Web Framework
fastapi>=0.104.0  # For Python 3.12 compatibility
uvicorn>=0.23.2
requests>=2.31.0
python-multipart>=0.0.6
pydantic>=2.4.0  # For Python 3.12 compatibility
typing_extensions>=4.7.0  # For Python 3.12 compatibility